+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-google-client-v1.2.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.1...@adobe/spacecat-shared-google-client-v1.2.2) (2024-08-26)
+
+
+### Bug Fixes
+
+* authentication token refresh ([#344](https://github.com/adobe/spacecat-shared/issues/344)) ([37cb442](https://github.com/adobe/spacecat-shared/commit/37cb442c7837241780af3fede345db6be66e72e2))
+
+# [@adobe/spacecat-shared-google-client-v1.2.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.2.0...@adobe/spacecat-shared-google-client-v1.2.1) (2024-08-24)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#345](https://github.com/adobe/spacecat-shared/issues/345)) ([ae2c7a6](https://github.com/adobe/spacecat-shared/commit/ae2c7a6104394a53d74f5c19465d6741751576ce))
+
 # [@adobe/spacecat-shared-google-client-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.19...@adobe/spacecat-shared-google-client-v1.2.0) (2024-08-20)