+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-utils-v1.19.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.19.5...@adobe/spacecat-shared-utils-v1.19.6) (2024-08-24)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#345](https://github.com/adobe/spacecat-shared/issues/345)) ([ae2c7a6](https://github.com/adobe/spacecat-shared/commit/ae2c7a6104394a53d74f5c19465d6741751576ce))
+
 # [@adobe/spacecat-shared-utils-v1.19.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.19.4...@adobe/spacecat-shared-utils-v1.19.5) (2024-08-20)